// src/api/generated/models/TokenHolderDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PriceDto } from './PriceDto';
export type TokenHolderDto = {
    id: number;
    address: string;
    balance: string;
    balance_data: PriceDto;
    created_at: string;
};